Separates code better.
