+++ react-native-customer-frontend/businesses/app/YourAppointmentsPage.tsx
@@ -4,24 +4,14 @@ import { Menu } from "@/components/Menu/Menu";
 import Styles from "@/constants/Styles";
 import { AppPageLink } from "@/components/App Page Link/AppPageLink";
 import { MenuItem } from "@/components/Menu Item/MenuItem";
-import { getLocaleDateTimeString, nextMidnightFrom } from "@/scripts/formatting";
-import { getAppointments, getAppointmentsForSpecificDay, getAppointmentsOnOrAfter } from "@/networking/controllers/customerController";
+import { getLocaleDateTimeString } from "@/scripts/formatting";
+import { getLaterAppointments, getTodaysAppointments } from "@/networking/controllers/customerController";
 
 export const YourAppointmentsPage = ({ navigation, route }: any) => {
 
     const TEST_CUSTOMER_ID = "66b84b5f4d019d6b83778176";
 
     const customerId = TEST_CUSTOMER_ID;
-    const LATER_APPOINTMENTS_ITEM_LIMIT = 3;
-    const LATER_APPOINTMENTS_SORT_BY = "startDateTime";
-    const LATER_APPOINTMENTS_SORT_ORDER = "asc";
-
-    const LATER_APPOINTMENTS_FILTERS = {
-        onOrAfter: nextMidnightFrom(Date.now()),
-        sortBy: LATER_APPOINTMENTS_SORT_BY,
-        sortOrder: LATER_APPOINTMENTS_SORT_ORDER,
-        limit: LATER_APPOINTMENTS_ITEM_LIMIT
-    }
 
     const [todaysAppointments, setTodaysAppointments] = useState([]);
 
@@ -32,9 +22,9 @@ export const YourAppointmentsPage = ({ navigation, route }: any) => {
     const [error, setError] = useState(false);
 
     useEffect(() => {
-        getAppointmentsForSpecificDay(customerId, Date.now()) // sort these by startDateTime too
+        getTodaysAppointments(customerId) // sort these by startDateTime too
         .then(json => setTodaysAppointments(json))
-        .then(() => getAppointments(customerId, LATER_APPOINTMENTS_FILTERS))
+        .then(() => getLaterAppointments(customerId))
         .then(json => setLaterAppointments(json))
         .then(() => setLoading(false))
         .catch(error => { setError(true); console.log(error); })
