+++ react-native-customer-frontend/businesses/networking/controllers/customerController.tsx
@@ -1,13 +1,33 @@
-import { customersRoute, customerRoute, customerAppointmentsRoute, customerFirstUpcomingAppointmentRoute, customerAppointmentsForSpecificDayRoute, customerAppointmentsOnOrAfterRoute } from "../routers/customerRouter";
+import { nextMidnightFrom } from "@/scripts/formatting";
+import { customerAppointmentsRoute, customerFirstUpcomingAppointmentRoute, customerAppointmentsForSpecificDayRoute, customerAppointmentsOnOrAfterRoute } from "../routers/customerRouter";
+
+const LATER_APPOINTMENTS_ITEM_LIMIT = 3;
+const LATER_APPOINTMENTS_SORT_BY = "startDateTime";
+const LATER_APPOINTMENTS_SORT_ORDER = "asc";
 
 async function getFirstUpcomingAppointment(customerId: string, currentDateTime: number) {
     return await fetch(customerFirstUpcomingAppointmentRoute(customerId, currentDateTime), { method: "GET" })
         .then(res => res.status === 200 ? res.json() : null);
 }
 
+async function getTodaysAppointments(customerId: string) {
+    return await getAppointmentsForSpecificDay(customerId, Date.now());
+}
+
 async function getAppointmentsForSpecificDay(customerId: string, currentDateTime: number) {
     return await fetch(customerAppointmentsForSpecificDayRoute(customerId, currentDateTime), { method: "GET" })
         .then(res => res.status === 200 ? res.json() : null);
+}
+
+async function getLaterAppointments(customerId: string) {
+    const LATER_APPOINTMENTS_FILTERS = {
+        onOrAfter: nextMidnightFrom(Date.now()),
+        sortBy: LATER_APPOINTMENTS_SORT_BY,
+        sortOrder: LATER_APPOINTMENTS_SORT_ORDER,
+        limit: LATER_APPOINTMENTS_ITEM_LIMIT
+    }
+
+    return await getAppointments(customerId, LATER_APPOINTMENTS_FILTERS);
 }
 
 async function getAppointmentsOnOrAfter(customerId: string, currentDateTime: number) {
@@ -17,7 +37,9 @@ async function getAppointmentsOnOrAfter(customerId: string, currentDateTime: num
 
 /*  Parameters:
  *  - onOrAfter: number
+ *  - after: number
  *  - onOrBefore: number
+ *  - before: number
  *  - limit: number
  *  - sort: { attribute: string, order: string }
  */
@@ -26,4 +48,4 @@ async function getAppointments(customerId: string, queryParams: object) {
         .then(res => res.status === 200 ? res.json() : null);
 }
 
-export { getFirstUpcomingAppointment, getAppointmentsForSpecificDay, getAppointmentsOnOrAfter, getAppointments };
+export { getFirstUpcomingAppointment, getTodaysAppointments, getLaterAppointments, getAppointmentsForSpecificDay, getAppointmentsOnOrAfter, getAppointments };
